Only redraw the store if you bought something
When selecting an item, then selecting purchase from the context menu, the player may end up buying something.  If so, we want to redraw the store.  If the player didn't buy something, redrawing the store is unecessary, because the inventory hasn't changed, and it may overwrite a message explaining why a purchase attempt was unsuccessful.