SDL2: refactor event handling and change drawing for menu items
Intended as preparation for adding a font selection dialog like the SDL front end now has.  The changes to appearance reduce the role of color in the menus, so as long as the colors at COLOUR_WHITE is distinct from the colors at COLOUR_SHADE and COLOUR_DARK, the menus should be usable.  That resolves https://github.com/angband/angband/issues/5596 .  Allows keyboard navigation of the menus (to avoid conflicts with user keymaps, no key is assigned for transferring focus to a menu; such keys can be configured with the "Menu Shortcuts" menu item).  Should allow mouse wheel events to change the integer menu entries (font and tile size).  Adds entries to sdl2init.txt so copies of that file generated by this version can not, without editing, be used by previous versions.