Handle monster groups when a ghost is skipped over in chunk_copy()
Avoids a assertion failure copying over the monster groups when both chunks have a ghost.

An alternate implementation of this, which would assume less about the internals of monster_remove_from_groups(), would copy over the ghost to skip (but still not overwrite dest->ghost) and then call delete_monster_idx() for the ghost to skip at the end of chunk_copy().  That would require modifying delete_monster_idx() to take a chunk argument, and that modifcation to delete_monster_idx() should be done in Vanilla to avoid unnecessary differences between FAAngband's code and Vanilla's.