docs: correct some details for blackguards and shooting/throwing
Match blackguards' adjustments for hit die and skills to what is currently in class.txt.  Throwing skill is separate from shooting though only rogues, rangers, and blackguards will have different values for those two skills.  Shooting and throwing skills do not affect range and only affect the damage for a hit when using the O-combat option (because the to-hit chance affects the critical hits with that option).