Allow switching between equipment/inventory/quiver/floor by mouse
To do so click the row above where an item menu is displayed in the main window (for instance the row saying "Select Item: (Equip ...)").  The location clicked must not be to the left of column with the selection letters or numbers for the menu.  Clicking in such a location dismisses the list regardless of the row clicked.  A context menu will appear which allows selection of the new source of the list or to back out from displaying the item menu.