Fold applying ego activation into ego_apply_magic()
That way, other paths which might create ego items (the debugging commands, birth items, the BRAND_WEAPON effect) also get the activation.  Also affects the generated dummy item for the ego knowledge menu, but, since the activation information is not displayed when OINFO_EGO is used, the player will not see a difference there.