Rework how keymaps are handled when extracting a direction
Only recognize a keymap trigger if its action is one walk or run command.  Also, depending on the context, allow a keymap trigger whose first key in the action is ESCAPE to break out of the selection process.  The latter resolves https://github.com/angband/angband/issues/6297 .  Such a keymap trigger can also break out of item, spell, curse, or effect selection.  They do not affect retrieving the input when the player is prompted for a string, quantity, confirmation, or single character.  Menus used in the knowledge or preferences menu are not affected unless they use the item/spell/curse/effect input methods.