cmake: use lower case for command invocations
Also use it for the first argument to function() and macro().  Matches what CMake prefers in its documentation and tutorials.  Remove whitespace between a command and the opening parenthesis.  Also correct some typos in comments and remove a DEPENDS argument from an add_custom_command(TARGET ...) which triggered a policy warning on OpenBSD.